Move corresponding files to the presentation layer package
