+++ app/src/main/java/com/simply/birthdayapp/presentation/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.simply.birthdayapp.ui.theme
+package com.simply.birthdayapp.presentation.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
